Split _plotWin.py’s export/refresh responsibilities, fix/pin mypy so CI and local dev agree, then add targeted tests around worker/data-transform/export paths.
