+++ src/qplot/__main__.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 from PyQt6 import QtWidgets as qtw
 
-from qplot._version import package_version
+from qplot._metadata import package_version
 from qplot.diagnostics import (
     configure_logging,
     install_excepthook,
